requirements.txt
bidict==0.21.2
click==7.1.2
dnspython==1.16.0
eventlet==0.30.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-SocketIO==5.0.1
greenlet==0.4.17
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.2
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.3
six==1.15.0
Werkzeug==1.0.1
